+++ test/phoenix/live_dashboard/pages/ecto_stats_page_test.exs
@@ -41,16 +41,14 @@ defmodule Phoenix.LiveDashboard.EctoStatsPageTest do
     assert rendered =~ "Phoenix.LiveDashboardTest.Repo"
     assert rendered =~ "Phoenix.LiveDashboardTest.PGRepo"
 
-    mysql_started? = start_mysql_repo()
+    start_mysql_repo!()
 
-    if mysql_started? do
-      {:ok, live, _} = live(build_conn(), ecto_stats_path())
-      rendered = render(live)
+    {:ok, live, _} = live(build_conn(), ecto_stats_path())
+    rendered = render(live)
 
-      assert rendered =~ "Phoenix.LiveDashboardTest.Repo"
-      assert rendered =~ "Phoenix.LiveDashboardTest.PGRepo"
-      assert rendered =~ "Phoenix.LiveDashboardTest.MySQLRepo"
-    end
+    assert rendered =~ "Phoenix.LiveDashboardTest.Repo"
+    assert rendered =~ "Phoenix.LiveDashboardTest.PGRepo"
+    assert rendered =~ "Phoenix.LiveDashboardTest.MySQLRepo"
 
     start_sqlite_repo!()
 
@@ -59,7 +57,7 @@ defmodule Phoenix.LiveDashboard.EctoStatsPageTest do
 
     assert rendered =~ "Phoenix.LiveDashboardTest.Repo"
     assert rendered =~ "Phoenix.LiveDashboardTest.PGRepo"
-    if mysql_started?, do: assert(rendered =~ "Phoenix.LiveDashboardTest.MySQLRepo")
+    assert rendered =~ "Phoenix.LiveDashboardTest.MySQLRepo"
     assert rendered =~ "Phoenix.LiveDashboardTest.SQLiteRepo"
   end
 
@@ -90,10 +88,10 @@ defmodule Phoenix.LiveDashboard.EctoStatsPageTest do
       assert {:ok, _, _} = live(build_conn(), ecto_stats_path(nav))
     end
 
-    if start_mysql_repo() do
-      for {nav, _} <- EctoMySQLExtras.queries(MySQLRepo) do
-        assert {:ok, _, _} = live(build_conn(), ecto_stats_path(nav))
-      end
+    start_mysql_repo!()
+
+    for {nav, _} <- EctoMySQLExtras.queries(MySQLRepo) do
+      assert {:ok, _, _} = live(build_conn(), ecto_stats_path(nav))
     end
 
     start_pg_repo!()
@@ -138,27 +136,27 @@ defmodule Phoenix.LiveDashboard.EctoStatsPageTest do
     refute rendered =~ "fuzzystrmatch"
     assert rendered =~ "hstore"
 
-    if start_mysql_repo() do
-      {:ok, live, _} = live(build_conn(), ecto_stats_path(:plugins, "", MySQLRepo))
+    start_mysql_repo!()
 
-      rendered = render(live)
-      assert rendered =~ "Version"
-      assert rendered =~ "Status"
-      assert rendered =~ "PERFORMANCE_SCHEMA"
-      assert rendered =~ "InnoDB"
+    {:ok, live, _} = live(build_conn(), ecto_stats_path(:plugins, "", MySQLRepo))
 
-      {:ok, live, _} =
-        live(
-          build_conn(),
-          ecto_stats_path(:plugins, "InnoDB", MySQLRepo)
-        )
+    rendered = render(live)
+    assert rendered =~ "Version"
+    assert rendered =~ "Status"
+    assert rendered =~ "PERFORMANCE_SCHEMA"
+    assert rendered =~ "InnoDB"
 
-      rendered = render(live)
-      assert rendered =~ "Version"
-      assert rendered =~ "Status"
-      refute rendered =~ "PERFORMANCE_SCHEMA"
-      assert rendered =~ "InnoDB"
-    end
+    {:ok, live, _} =
+      live(
+        build_conn(),
+        ecto_stats_path(:plugins, "InnoDB", MySQLRepo)
+      )
+
+    rendered = render(live)
+    assert rendered =~ "Version"
+    assert rendered =~ "Status"
+    refute rendered =~ "PERFORMANCE_SCHEMA"
+    assert rendered =~ "InnoDB"
 
     start_sqlite_repo!()
 
@@ -204,24 +202,8 @@ defmodule Phoenix.LiveDashboard.EctoStatsPageTest do
     start_supervised!(PGRepo)
   end
 
-  defp start_mysql_repo do
-    with mysql_url when is_binary(mysql_url) <- System.get_env("MYSQL_URL"),
-         {:ok, _} <- start_supervised(MySQLRepo),
-         {:ok, _} <- mysql_query() do
-      true
-    else
-      _ ->
-        stop_supervised(MySQLRepo)
-        false
-    end
-  end
-
-  defp mysql_query do
-    try do
-      MySQLRepo.query("SELECT 1", [], timeout: 500, pool_timeout: 500)
-    catch
-      :exit, reason -> {:error, reason}
-    end
+  defp start_mysql_repo! do
+    start_supervised!(MySQLRepo)
   end
 
   defp start_sqlite_repo! do